MonitoringForUpdateStratagia/requirements.txt
apscheduler==3.9.1
aiohttp==3.10.11
python-telegram-bot==20.0
requests==2.32.2
python-dotenv==1.0.0
pytz==2023.3
dotenv==4.1.0